refactor: refactor workflows titles and lables
