--- .github/workflows/_called_publish-report-discord.yml
@@ -1,82 +1,0 @@
-name: Publish reports in Discord
-
-on:
-  workflow_call:
-    inputs:
-
-      app-build-pub-in-discord:
-        required: true
-        type: boolean
-      app-build-status:
-        required: false
-        type: boolean
-
-      app-sb-deploy-pub-report-in-discord:
-        required: true
-        type: boolean
-      app-sb-deploy-status:
-        required: false
-        type: boolean
-
-      app-unit-test-pub-report-in-discord:
-        required: false
-        type: boolean
-      app-unit-test-status:
-        required: false
-        type: boolean
-      app-unit-test-codecov-percentage:
-        required: false
-        type: string
-      app-unit-test-codecov-diff:
-        required: false
-        type: string
-
-
-    secrets:
-#      gh_token:
-#        required: true
-      gh_pages_full_branch:
-        required: true
-      discord_alert_ui_web_hook:
-        required: true
-
-jobs:
-
-  publish-summary-report-discord:
-    name: Publish summary report to Discord
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-
-      - uses: actions/github-script@v6
-        id: discord_embed_report
-        env:
-          IS_APP_SB_BUILD_REPORT: ${{ inputs.app-build-pub-in-discord }}
-          IS_APP_UNIT_TEST_REPORT: ${{ inputs.app-unit-test-pub-report-in-discord }}
-          IS_APP_E2E_TEST_REPORT: false
-          IS_SB_UNIT_TEST_REPORT: false
-          IS_SB_E2E_TEST_REPORT: false
-          IS_APP_SB_DEPLOYMENT_REPORT: ${{ inputs.app-sb-deploy-pub-report-in-discord }}
-
-          APP_UNIT_TEST_PERCENTAGE: ${{ inputs.app-unit-test-codecov-percentage }}
-          APP_UNIT_TEST_DIFF: ${{ inputs.app-unit-test-codecov-diff }}
-          APP_BUILD_STATUS: ${{ inputs.app-build-status }}
-          APP_UNIT_TEST_STATUS: ${{ inputs.app-unit-test-status }}
-
-          APP_DEPLOYMENT_STATUS: ${{ inputs.app-sb-deploy-status }}
-
-          GITHUB_HEAD_REF: ${{ secrets.GITHUB_HEAD_REF }}
-          GITHUB_REF: ${{ secrets.GITHUB_REF }}
-          GITHUB_REF_NAME: ${{ secrets.GITHUB_REF_NAME }}
-        with:
-          result-encoding: string
-          script: |
-            const script = require('./scripts/ci/github-script-src/create-discord-report-embed-body.js')
-            return await script({github, context, core})
-
-      - name: Discord notification
-        env:
-          DISCORD_WEBHOOK: ${{ secrets.DISCORD_ALERT_UI_WEB_HOOK }}
-          DISCORD_USERNAME: Basilisk-UI GH Bot
-          DISCORD_EMBEDS: ${{ steps.discord_embed_report.outputs.result }}
-        uses: Ilshidur/action-discord@master
