+++ .github/workflows/_called_report-status-in-discord.yml
@@ -1,24 +1,17 @@
-name: Publish reports in issue comment
+name: Reports in Discord
 
 on:
   workflow_call:
     inputs:
-      publish-artifacts-list:
+
+      app-storybook-build-pub-report:
         required: true
         type: boolean
-
-      report-msg-title:
-        required: false
-        type: string
-
-      app-build-pub-report:
-        required: true
-        type: boolean
-      app-build-status:
+      app-storybook-build-status:
         required: false
         type: boolean
 
-      app-sb-deploy-pub-report:
+      app-storybook-deploy-pub-report:
         required: true
         type: boolean
       app-sb-deploy-status:
@@ -39,47 +32,48 @@ on:
         type: string
 
     secrets:
-      gh_token:
-        required: true
       gh_pages_full_branch:
         required: true
-      gh_pages_custom_domain:
+      discord_alert_ui_web_hook:
         required: true
 
 jobs:
-  publish-report-issue-comment:
-    name: Publish issue comment
+  publish-report-in-discord:
+    name: Publish
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
 
-      - uses: actions/github-script@v6
-        id: issue-comment-data
+      - name: Prepare report content
+        uses: actions/github-script@v6
+        id: discord_embed_report
         env:
-
-          REPORT_MSG_TITLE: ${{inputs.report-msg-title}}
-          PUBLISH_ARTIFACTS_LIST: ${{inputs.publish-artifacts-list}}
-          PUBLISH_ARTIFACTS_WORKFLOW_DISPATCH_FILE: wfd_publish-issue-comment-with-artifacts.yml
-
-          IS_APP_SB_BUILD_REPORT: ${{ inputs.app-build-pub-report }}
+          IS_APP_SB_BUILD_REPORT: ${{ inputs.app-storybook-build-pub-report }}
           IS_APP_UNIT_TEST_REPORT: ${{ inputs.app-unit-test-pub-report }}
           IS_APP_E2E_TEST_REPORT: false
           IS_SB_UNIT_TEST_REPORT: false
           IS_SB_E2E_TEST_REPORT: false
-          IS_APP_SB_DEPLOYMENT_REPORT: ${{ inputs.app-sb-deploy-pub-report }}
+          IS_APP_SB_DEPLOYMENT_REPORT: ${{ inputs.app-storybook-deploy-pub-report }}
 
           APP_UNIT_TEST_PERCENTAGE: ${{ inputs.app-unit-test-codecov-percentage }}
           APP_UNIT_TEST_DIFF: ${{ inputs.app-unit-test-codecov-diff }}
-          APP_BUILD_STATUS: ${{ inputs.app-build-status }}
+          APP_STORYBOOK_BUILD_STATUS: ${{ inputs.app-storybook-build-status }}
           APP_UNIT_TEST_STATUS: ${{ inputs.app-unit-test-status }}
 
           APP_DEPLOYMENT_STATUS: ${{ inputs.app-sb-deploy-status }}
 
-          GITHUB_TOKEN: ${{ secrets.GH_TOKEN }}
-          GH_TOKEN: ${{ secrets.GH_TOKEN }}
-          GH_PAGES_CUSTOM_DOMAIN: ${{ secrets.gh_pages_custom_domain }}
-
+          GITHUB_HEAD_REF: ${{ secrets.GITHUB_HEAD_REF }}
+          GITHUB_REF: ${{ secrets.GITHUB_REF }}
+          GITHUB_REF_NAME: ${{ secrets.GITHUB_REF_NAME }}
         with:
+          result-encoding: string
           script: |
-            const script = require('./scripts/ci/github-script-src/publish-issue-comment.js')
+            const script = require('./scripts/ci/github-script-src/create-discord-report-embed-body.js')
             return await script({github, context, core})
+
+      - name: Publish Discord message
+        env:
+          DISCORD_WEBHOOK: ${{ secrets.DISCORD_ALERT_UI_WEB_HOOK }}
+          DISCORD_USERNAME: Basilisk-UI GH Bot
+          DISCORD_EMBEDS: ${{ steps.discord_embed_report.outputs.result }}
+        uses: Ilshidur/action-discord@master
