+++ .github/workflows/_called_run-e2e-tests-app.yml
@@ -1,4 +1,4 @@
-name: App E2E tests
+name: E2E tests :: App
 
 on:
   workflow_call:
@@ -11,11 +11,12 @@ on:
         type: string
 
 jobs:
-  run_app_e2e_tests:
-    name: Run tests
+  run:
+    name: Run
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/setup-node@v2
+      - name: Setup Node.js
+        uses: actions/setup-node@v2
         with:
           node-version: 17.3
