+++ .github/workflows/wf_build-deploy-app-storybook.yml
@@ -1,4 +1,4 @@
-name: Build Deploy App and Storybook
+name: Build-Deploy-Report
 on:
   push:
     branches:
@@ -12,19 +12,19 @@ on:
       - develop
 
 jobs:
-  build-app-storybook:
-    name: Build App and Storybook
-    uses: ./.github/workflows/_called_build-app-and-storybook.yml
+  build:
+    name: Build
+    uses: ./.github/workflows/_called_build.yml
     with:
       app-build-artifact-name: app-build-files
       storybook-build-artifact-name: sb-build-files
       app-node-modules-cache-key: cache-node-modules-ui-app
 
-  deploy-app-storybook-builds:
-    name: Deploy App&Storybook builds
-    needs: build-app-storybook
-    uses: ./.github/workflows/_called_deploy-app-and-storybook.yml
-    if: ${{ needs.build-app-storybook.result=='success' }}
+  deploy:
+    name: Deploy
+    needs: build
+    uses: ./.github/workflows/_called_deploy.yml
+    if: ${{ needs.build.result=='success' }}
     with:
       app-build-artifact-name: app-build-files
       storybook-build-artifact-name: sb-build-files
@@ -32,20 +32,20 @@ jobs:
       gh_token: ${{ secrets.GH_TOKEN }}
       gh_pages_full_branch: ${{ secrets.GH_PAGES_FULL_BRANCH }}
 
-  report-statuses-issue-comment:
-    name: Reportr issue comment
-    uses: ./.github/workflows/_called_publish-report-issue-comment.yml
+  report-statuses-in-issue:
+    name: Report in issue
+    uses: ./.github/workflows/_called_report-status-in-issue.yml
     if: ${{ always() }}
-    needs: [ build-app-storybook, deploy-app-storybook-builds ]
+    needs: [ build, deploy ]
     with:
       report-msg-title: Basilisk-UI build reporter
       publish-artifacts-list: true
 
-      app-build-pub-report: true
-      app-build-status: ${{ needs.build-app-storybook.result == 'success' }}
+      app-storybook-build-pub-report: true
+      app-storybook-build-status: ${{ needs.build.result == 'success' }}
 
       app-sb-deploy-pub-report: true
-      app-sb-deploy-status: ${{ needs.deploy-app-storybook-builds.result == 'success' }}
+      app-sb-deploy-status: ${{ needs.deploy.result == 'success' }}
 
       app-unit-test-pub-report: false
     secrets:
