+++ scripts/ci/github-script-src/discord-reporter-components/app-storybpook-deployment-report.js
@@ -1,7 +1,4 @@
 module.exports = ({ APP_DEPLOYMENT_STATUS, context }) => {
-  const workingBranch = context.ref.replace('refs/heads/', '');
-  const repoUrl = context.payload.repository.html_url;
-
   const appSbDeploymentData = [
     {
       name: ':small_blue_diamond:  App / Storybook deploy',
